[AV2-1010] added styles related for events. update components to work with new UI. remove unwanted components. merge two event services to one service.
