+++ projects/v3/src/app/pages/events/event-list/event-list.module.ts
@@ -1,18 +1,18 @@
-import { SharedModule } from '@shared/shared.module';
+import { ComponentsModule } from '@v3/components/components.module';
 import { RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
 import { EventListComponent } from './event-list.component';
-import { EventListService } from './event-list.service';
+import { EventService } from '@v3/services/event.service';
 
 @NgModule({
   imports: [
-    SharedModule,
+    ComponentsModule,
   ],
   declarations: [
     EventListComponent
   ],
   providers: [
-    EventListService
+    EventService
   ],
   exports: [
     EventListComponent
