+++ projects/v3/src/app/pages/events/events.page.scss
@@ -2,3 +2,10 @@
   font-size: 200px;
   margin-top: 20%;
 }
+
+ion-grid {
+  height: -webkit-fill-available;
+  ion-row {
+    height: -webkit-fill-available;
+  }
+}
