--- projects/v3/src/app/pages/events/events/events-routing.component.ts
@@ -1,6 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  template: '<ion-router-outlet></ion-router-outlet>'
-})
-export class EventsRoutingComponent {}
