--- projects/v3/src/app/pages/events/events/events.component.scss
@@ -1,4 +1,0 @@
-#calendar-icon {
-  font-size: 200px;
-  margin-top: 20%;
-}
