+++ projects/v3/src/app/pages/v3/v3.page.ts
@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 import { ModalController, PopoverController } from '@ionic/angular';
 import { AnimationsService } from '@v3/services/animations.service';
 import { SettingsPage } from '../settings/settings.page';
-import { SharedService } from '@v3/services/shared.service';
 
 @Component({
   selector: 'app-v3',
@@ -35,11 +34,8 @@ export class V3Page {
   constructor(
     private modalController: ModalController,
     private popoverController: PopoverController,
-    private animationService: AnimationsService,
-    private sharedService: SharedService
+    private animationService: AnimationsService
   ) {
-    // TODO need to move to experience page to call before navigate to experience
-    this.sharedService.initWebServices();
   }
 
   async presentModal() {
